+++ processor/src/test/fixtures/bad_input/com/example/injector/EnclosingInterfaceInjectorModel.java
@@ -1,0 +1,19 @@
+package com.example.injector;
+
+import sting.Injectable;
+import sting.Injector;
+
+@Injector
+interface EnclosingInterfaceInjectorModel
+{
+  MyModel getMyModel();
+
+  @Injectable
+  class MyModel
+  {
+  }
+
+  interface Foo
+  {
+  }
+}